Move all MeTTa related modules into crate::metta
