+++ lib/src/lib.rs
@@ -3,13 +3,6 @@ extern crate mopa;
 
 pub mod atom;
 pub mod space;
-
-pub mod text;
-pub mod common;
-pub mod interpreter;
-pub mod arithmetics;
-
-mod types;
-mod examples;
+pub mod metta;
 
 pub use atom::*;
