+++ lib/src/metta/common/mod.rs
@@ -1,4 +1,8 @@
-use super::*;
+mod arithmetics;
+
+pub use arithmetics::*;
+
+use crate::*;
 use std::cell::RefCell;
 
 use std::fmt::Debug;
