+++ lib/src/metta/examples/mod.rs
