+++ lib/src/metta/mod.rs
@@ -1,0 +1,6 @@
+pub mod text;
+pub mod interpreter;
+
+mod types;
+mod common;
+mod examples;
